Move storage defaults to storage.ts to fix test failure
